# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API for job posting website",
        "description": "API for website that provides job vacancies in various IT companies all over Belarus. This web-application is ment for students as well as for people with no (or little) working experience in IT sphere, as it offers exactly those positions, where no experience is required.",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/user/password_reset": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Resets password for user",
                "description": "Resets password for user in case it was forgotten",
                "operationId": "password_reset",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email of the user",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset attempt received",
                        "examples": {
                            "application/json": {
                                "data": "If email is correct, you will get the letter with your new password within few minutes"
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid email",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": "422",
                                    "title": "Invalid data",
                                    "description": {
                                        "email": "[The email must be a valid email address.]"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Registers new user",
                "description": "Registers new user on website",
                "operationId": "registerUser",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name of the user",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email of the user",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password of the user",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Password confirmation",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "City where the user live",
                        "required": true,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Mobile phone number of the user",
                        "required": false,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "newsletter",
                        "in": "formData",
                        "description": "Agreement of the user for getting newsletter with suitable jobs (filter by city)",
                        "required": true,
                        "type": "string",
                        "default": "0"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "examples": {
                            "application/json": {
                                "data": "user registered"
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid user data",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": "422",
                                    "title": "Invalid data",
                                    "description": {
                                        "email": "[The email has already been taken.]",
                                        "city": "[The city field is required.]"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/show": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Show user info",
                "description": "Return information about currently authenticated user",
                "operationId": "showUser",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "User information fetched successfully",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "id": "32",
                                    "name": "Alex",
                                    "email": "alex@mailserver.org",
                                    "city": "\u041c\u0438\u043d\u0441\u043a",
                                    "phone": "+375123456789",
                                    "newsletter": "1",
                                    "status": "1",
                                    "created_at": "2017-07-25 12:13:36",
                                    "updated_at": "2017-07-27 10:57:36"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token absent",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": 400,
                                    "title": "Token absent",
                                    "detail": "Token can not be parsed from the request headers"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/user/update": {
            "patch": {
                "tags": [
                    "user"
                ],
                "summary": "Update user info",
                "description": "Update information about currently authenticated user",
                "operationId": "updateUser",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name of the user",
                        "required": false,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password of the user",
                        "required": false,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Password confirmation (REQUIRED IF password IS SET)",
                        "required": false,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "city",
                        "in": "formData",
                        "description": "City where the user live",
                        "required": false,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Mobile phone number of the user",
                        "required": false,
                        "type": "string",
                        "maxLength": 255
                    },
                    {
                        "name": "newsletter",
                        "in": "formData",
                        "description": "Agreement of the user for getting newsletter with suitable jobs",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information updated successfully",
                        "examples": {
                            "application/json": {
                                "data": {
                                    "id": "32",
                                    "name": "Alex",
                                    "email": "alex@mailserver.org",
                                    "city": "\u041c\u0438\u043d\u0441\u043a",
                                    "phone": "+375987654321",
                                    "newsletter": "1",
                                    "status": "1",
                                    "created_at": "2017-07-25 12:13:36",
                                    "updated_at": "2017-07-27 10:57:36"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Provided data is invalid and can not be used",
                        "examples": {
                            "application/json": {
                                "errors": {
                                    "status": 422,
                                    "title": "Invalid data",
                                    "detail": {
                                        "name": "The name field is required."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}